miner/requirements.txt
biopython>=1.83
pyyaml>=6.0
pandas>=2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
